Refactor proving
