+++ crates/prover/src/da_block_handler.rs
@@ -8,38 +8,36 @@ use anyhow::anyhow;
 use bitcoin_da::helpers::compression::compress_blob;
 use borsh::{BorshDeserialize, BorshSerialize};
 use citrea_common::cache::L1BlockCache;
-use citrea_common::da::{extract_sorted_sequencer_commitments, get_da_block_at_height};
-use citrea_common::utils::{
-    check_l2_range_exists, filter_out_proven_commitments, merge_state_diffs,
-};
+use citrea_common::da::get_da_block_at_height;
+use citrea_common::utils::merge_state_diffs;
 use citrea_common::ProverConfig;
 use citrea_primitives::MAX_TXBODY_SIZE;
 use rand::Rng;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use sov_db::ledger_db::ProverLedgerOps;
-use sov_db::schema::types::{BatchNumber, SlotNumber, StoredProof, StoredStateTransition};
-use sov_modules_api::{BlobReaderTrait, DaSpec, StateDiff, Zkvm};
+use sov_db::schema::types::{BatchNumber, SlotNumber};
+use sov_modules_api::{DaSpec, StateDiff, Zkvm};
 use sov_rollup_interface::da::{BlockHeaderTrait, SequencerCommitment};
-use sov_rollup_interface::rpc::SoftConfirmationStatus;
 use sov_rollup_interface::services::da::{DaService, SlotData};
 use sov_rollup_interface::soft_confirmation::SignedSoftConfirmation;
 use sov_rollup_interface::spec::SpecId;
-use sov_rollup_interface::zk::{Proof, StateTransitionData, ZkvmHost};
+use sov_rollup_interface::zk::ZkvmHost;
 use sov_stf_runner::ProverService;
 use tokio::select;
 use tokio::sync::{mpsc, Mutex};
 use tokio::time::{sleep, Duration};
 use tokio_util::sync::CancellationToken;
-use tracing::{error, info, warn};
+use tracing::{error, info};
+
+use crate::errors::L1ProcessingError;
+use crate::proving::{data_to_prove, extract_and_store_proof, prove_l1};
 
 type CommitmentStateTransitionData<Witness, Da> = (
     VecDeque<Vec<Witness>>,
     VecDeque<Vec<SignedSoftConfirmation>>,
     VecDeque<Vec<<<Da as DaService>::Spec as DaSpec>::BlockHeader>>,
 );
-
-type TxId<Da> = <Da as DaService>::TransactionId;
 
 pub(crate) struct L1BlockHandler<Vm, Da, Ps, DB, StateRoot, Witness>
 where
@@ -75,7 +73,7 @@ where
     Da: DaService,
     Vm: ZkvmHost + Zkvm,
     Ps: ProverService<Vm, DaService = Da, StateRoot = StateRoot, Witness = Witness>,
-    DB: ProverLedgerOps + Clone,
+    DB: ProverLedgerOps + Clone + 'static,
     StateRoot: BorshDeserialize
         + BorshSerialize
         + Serialize
@@ -172,35 +170,63 @@ where
                 )
                 .unwrap();
 
-            let mut da_data: Vec<<<Da as DaService>::Spec as DaSpec>::BlobTransaction> =
-                self.da_service.extract_relevant_blobs(l1_block);
+            let data_to_prove = data_to_prove(
+                self.da_service.clone(),
+                self.ledger_db.clone(),
+                self.sequencer_pub_key.clone(),
+                self.sequencer_da_pub_key.clone(),
+                self.l1_block_cache.clone(),
+                l1_block.clone(),
+                Some(true),
+            );
 
-            // if we don't do this, the zk circuit can't read the sequencer commitments
-            da_data.iter_mut().for_each(|blob| {
-                blob.full_data();
-            });
+            let (sequencer_commitments, state_transitions) = match data_to_prove.await {
+                Ok((commitments, transitions)) => (commitments, transitions),
+                Err(e) => match e {
+                    L1ProcessingError::NoSeqCommitments { l1_height } => {
+                        info!("No sequencer commitment found at height {}", l1_height,);
+                        self.ledger_db
+                                .set_last_scanned_l1_height(SlotNumber(l1_height))
+                                .unwrap_or_else(|_| {
+                                    panic!(
+                                        "Failed to put prover last scanned l1 height in the ledger db {}",
+                                        l1_height
+                                    )
+                                });
 
-            let sequencer_commitments: Vec<SequencerCommitment> =
-                extract_sorted_sequencer_commitments::<Da>(
-                    self.da_service.clone(),
-                    l1_block.clone(),
-                    self.sequencer_da_pub_key.as_slice(),
-                );
-
-            if sequencer_commitments.is_empty() {
-                info!("No sequencer commitment found at height {}", l1_height,);
-                self.ledger_db
-                    .set_last_scanned_l1_height(SlotNumber(l1_height))
-                    .unwrap_or_else(|_| {
-                        panic!(
-                            "Failed to put prover last scanned l1 height in the ledger db {}",
+                        self.pending_l1_blocks.pop_front();
+                        continue;
+                    }
+                    L1ProcessingError::DuplicateCommitments { l1_height } => {
+                        info!(
+                            "All sequencer commitments are duplicates from a former DA block {}",
                             l1_height
-                        )
-                    });
+                        );
+                        self.ledger_db
+                            .set_last_scanned_l1_height(SlotNumber(l1_height))
+                            .unwrap_or_else(|_| {
+                                panic!(
+                                    "Failed to put prover last scanned l1 height in the ledger db {}",
+                                    l1_height
+                                )
+                            });
 
-                self.pending_l1_blocks.pop_front();
-                continue;
-            }
+                        self.pending_l1_blocks.pop_front();
+                        continue;
+                    }
+                    L1ProcessingError::L2RangeMissing {
+                        start_block_number,
+                        end_block_number,
+                    } => {
+                        error!("L2 range of commitments is not synced yet: {start_block_number} - {end_block_number}");
+                        break;
+                    }
+                    L1ProcessingError::Other(msg) => {
+                        error!("{msg}");
+                        return Err(anyhow!("{}", msg));
+                    }
+                },
+            };
 
             info!(
                 "Processing {} sequencer commitments at height {}",
@@ -208,97 +234,23 @@ where
                 l1_block.header().height(),
             );
 
-            // If the L2 range does not exist, we break off the local loop getting back to
-            // the outer loop / select to make room for other tasks to run.
-            // We retry the L1 block there as well.
-            if !check_l2_range_exists(
-                &self.ledger_db,
-                sequencer_commitments[0].l2_start_block_number,
-                sequencer_commitments[sequencer_commitments.len() - 1].l2_end_block_number,
-            ) {
-                break;
-            }
-
             // if proof_sampling_number is 0, then we always prove and submit
             // otherwise we submit and prove with a probability of 1/proof_sampling_number
             let should_prove = self.prover_config.proof_sampling_number == 0
                 || rand::thread_rng().gen_range(0..self.prover_config.proof_sampling_number) == 0;
 
-            let (sequencer_commitments, preproven_commitments) =
-                filter_out_proven_commitments(&self.ledger_db, &sequencer_commitments)?;
-
-            if sequencer_commitments.is_empty() {
-                info!(
-                    "All sequencer commitments are duplicates from a former DA block {}",
-                    l1_height
-                );
-                self.ledger_db
-                    .set_last_scanned_l1_height(SlotNumber(l1_height))
-                    .unwrap_or_else(|_| {
-                        panic!(
-                            "Failed to put prover last scanned l1 height in the ledger db {}",
-                            l1_height
-                        )
-                    });
-
-                self.pending_l1_blocks.pop_front();
-                continue;
-            }
-
-            let da_block_header_of_commitments: <<Da as DaService>::Spec as DaSpec>::BlockHeader =
-                l1_block.header().clone();
-
-            let hash = da_block_header_of_commitments.hash();
-
             if should_prove {
                 if l1_height >= self.skip_submission_until_l1 {
-                    let sequencer_commitments_groups = break_sequencer_commitments_into_groups(
-                        &self.ledger_db,
-                        &sequencer_commitments,
-                    )?;
-
-                    let submitted_proofs = self
-                        .ledger_db
-                        .get_proofs_by_l1_height(l1_height)?
-                        .unwrap_or(vec![]);
-                    for sequencer_commitment_range in sequencer_commitments_groups {
-                        // There is no ongoing bonsai session to recover
-                        let transition_data: StateTransitionData<StateRoot, Witness, Da::Spec> =
-                            self.create_state_transition_data(
-                                &sequencer_commitments,
-                                sequencer_commitment_range,
-                                &preproven_commitments,
-                                da_block_header_of_commitments.clone(),
-                                da_data.clone(),
-                                l1_block,
-                            )
-                            .await?;
-
-                        // check if transition data is already proven by crash recovery
-                        if !state_transition_already_proven::<StateRoot, Witness, Da>(
-                            &transition_data,
-                            &submitted_proofs,
-                        ) {
-                            let (tx_id, proof) =
-                                generate_and_submit_proof::<Ps, Vm, Da, StateRoot, Witness>(
-                                    self.prover_service.clone(),
-                                    self.da_service.clone(),
-                                    transition_data,
-                                    hash.clone(),
-                                )
-                                .await?;
-
-                            extract_and_store_proof::<DB, Da, Vm, StateRoot>(
-                                self.ledger_db.clone(),
-                                tx_id,
-                                proof,
-                                self.code_commitments_by_spec.clone(),
-                            )
-                            .await?;
-                        }
-
-                        save_commitments(self.ledger_db.clone(), &sequencer_commitments, l1_height);
-                    }
+                    prove_l1(
+                        self.da_service.clone(),
+                        self.prover_service.clone(),
+                        self.ledger_db.clone(),
+                        self.code_commitments_by_spec.clone(),
+                        l1_block.clone(),
+                        sequencer_commitments,
+                        state_transitions,
+                    )
+                    .await?;
                 } else {
                     info!("Skipping proving for l1 height {}", l1_height);
                 }
@@ -317,76 +269,6 @@ where
             self.pending_l1_blocks.pop_front();
         }
         Ok(())
-    }
-
-    async fn create_state_transition_data(
-        &self,
-        sequencer_commitments: &[SequencerCommitment],
-        sequencer_commitments_range: RangeInclusive<usize>,
-        preproven_commitments: &[usize],
-        da_block_header_of_commitments: <<Da as DaService>::Spec as DaSpec>::BlockHeader,
-        da_data: Vec<<<Da as DaService>::Spec as DaSpec>::BlobTransaction>,
-        l1_block: &Da::FilteredBlock,
-    ) -> Result<StateTransitionData<StateRoot, Witness, Da::Spec>, anyhow::Error> {
-        let first_l2_height_of_l1 =
-            sequencer_commitments[*sequencer_commitments_range.start()].l2_start_block_number;
-        let last_l2_height_of_l1 =
-            sequencer_commitments[*sequencer_commitments_range.end()].l2_end_block_number;
-        let (
-            state_transition_witnesses,
-            soft_confirmations,
-            da_block_headers_of_soft_confirmations,
-        ) = get_state_transition_data_from_commitments(
-            &sequencer_commitments[sequencer_commitments_range.clone()],
-            &self.da_service,
-            &self.ledger_db,
-            &self.l1_block_cache,
-        )
-        .await?;
-        let initial_state_root = self
-            .ledger_db
-            .get_l2_state_root::<StateRoot>(first_l2_height_of_l1 - 1)?
-            .expect("There should be a state root");
-        let initial_batch_hash = self
-            .ledger_db
-            .get_soft_confirmation_by_number(&BatchNumber(first_l2_height_of_l1))?
-            .ok_or(anyhow!(
-                "Could not find soft batch at height {}",
-                first_l2_height_of_l1
-            ))?
-            .prev_hash;
-
-        let final_state_root = self
-            .ledger_db
-            .get_l2_state_root::<StateRoot>(last_l2_height_of_l1)?
-            .expect("There should be a state root");
-
-        let (inclusion_proof, completeness_proof) = self
-            .da_service
-            .get_extraction_proof(l1_block, &da_data)
-            .await;
-
-        let transition_data: StateTransitionData<StateRoot, Witness, Da::Spec> =
-            StateTransitionData {
-                initial_state_root,
-                final_state_root,
-                initial_batch_hash,
-                da_data,
-                da_block_header_of_commitments,
-                inclusion_proof,
-                completeness_proof,
-                soft_confirmations,
-                state_transition_witnesses,
-                da_block_headers_of_soft_confirmations,
-                preproven_commitments: preproven_commitments.to_vec(),
-                sequencer_commitments_range: (
-                    *sequencer_commitments_range.start() as u32,
-                    *sequencer_commitments_range.end() as u32,
-                ),
-                sequencer_public_key: self.sequencer_pub_key.clone(),
-                sequencer_da_public_key: self.sequencer_da_pub_key.clone(),
-            };
-        Ok(transition_data)
     }
 
     async fn check_and_recover_ongoing_proving_sessions(&self) -> Result<(), anyhow::Error> {
@@ -591,149 +473,4 @@ pub(crate) fn break_sequencer_commitments_into_groups<DB: ProverLedgerOps>(
     // Add it anyway
     result_range.push(range);
     Ok(result_range)
-}
-
-pub(crate) async fn generate_and_submit_proof<Ps, Vm, Da, StateRoot, Witness>(
-    prover_service: Arc<Ps>,
-    da_service: Arc<Da>,
-    transition_data: StateTransitionData<StateRoot, Witness, Da::Spec>,
-    hash: <<Da as DaService>::Spec as DaSpec>::SlotHash,
-) -> Result<(TxId<Da>, Proof), anyhow::Error>
-where
-    Vm: ZkvmHost + Zkvm,
-    Ps: ProverService<Vm, DaService = Da, StateRoot = StateRoot, Witness = Witness>,
-    Da: DaService,
-{
-    prover_service.submit_witness(transition_data).await;
-
-    prover_service.prove(hash.clone()).await?;
-
-    prover_service
-        .wait_for_proving_and_send_to_da(hash.clone(), &da_service)
-        .await
-        .map_err(|e| anyhow!("Failed to prove and send to DA: {}", e))
-}
-
-pub(crate) fn state_transition_already_proven<StateRoot, Witness, Da>(
-    state_transition: &StateTransitionData<StateRoot, Witness, Da::Spec>,
-    proofs: &Vec<StoredProof>,
-) -> bool
-where
-    Da: DaService,
-    StateRoot: BorshDeserialize
-        + BorshSerialize
-        + Serialize
-        + DeserializeOwned
-        + Clone
-        + AsRef<[u8]>
-        + Debug,
-    Witness: Default + BorshDeserialize + Serialize + DeserializeOwned,
-{
-    for proof in proofs {
-        if proof.state_transition.initial_state_root == state_transition.initial_state_root.as_ref()
-            && proof.state_transition.final_state_root == state_transition.final_state_root.as_ref()
-            && proof.state_transition.sequencer_commitments_range
-                == state_transition.sequencer_commitments_range
-        {
-            return true;
-        }
-    }
-    false
-}
-
-pub(crate) async fn extract_and_store_proof<DB, Da, Vm, StateRoot>(
-    ledger_db: DB,
-    tx_id: <Da as DaService>::TransactionId,
-    proof: Proof,
-    code_commitments_by_spec: HashMap<SpecId, Vm::CodeCommitment>,
-) -> Result<(), anyhow::Error>
-where
-    Da: DaService,
-    DB: ProverLedgerOps,
-    Vm: ZkvmHost + Zkvm,
-    StateRoot: BorshDeserialize
-        + BorshSerialize
-        + Serialize
-        + DeserializeOwned
-        + Clone
-        + AsRef<[u8]>
-        + Debug,
-{
-    let tx_id_u8 = tx_id.into();
-
-    // l1_height => (tx_id, proof, transition_data)
-    // save proof along with tx id to db, should be queriable by slot number or slot hash
-    let transition_data = Vm::extract_output::<<Da as DaService>::Spec, StateRoot>(&proof)
-        .expect("Proof should be deserializable");
-
-    match &proof {
-        Proof::PublicInput(_) => {
-            warn!("Proof is public input, skipping");
-        }
-        Proof::Full(data) => {
-            info!("Verifying proof!");
-            let code_commitment = code_commitments_by_spec
-                .get(&transition_data.last_active_spec_id)
-                .expect("Proof public input must contain valid spec id");
-            Vm::verify(data, code_commitment)
-                .map_err(|err| anyhow!("Failed to verify proof: {:?}. Skipping it...", err))?;
-        }
-    }
-
-    info!("transition data: {:?}", transition_data);
-
-    let slot_hash = transition_data.da_slot_hash.into();
-
-    let stored_state_transition = StoredStateTransition {
-        initial_state_root: transition_data.initial_state_root.as_ref().to_vec(),
-        final_state_root: transition_data.final_state_root.as_ref().to_vec(),
-        state_diff: transition_data.state_diff,
-        da_slot_hash: slot_hash,
-        sequencer_commitments_range: transition_data.sequencer_commitments_range,
-        sequencer_public_key: transition_data.sequencer_public_key,
-        sequencer_da_public_key: transition_data.sequencer_da_public_key,
-        preproven_commitments: transition_data.preproven_commitments,
-        validity_condition: borsh::to_vec(&transition_data.validity_condition).unwrap(),
-    };
-    let l1_height = ledger_db
-        .get_l1_height_of_l1_hash(slot_hash)?
-        .expect("l1 height should exist");
-
-    if let Err(e) = ledger_db.insert_proof_data_by_l1_height(
-        l1_height,
-        tx_id_u8,
-        proof,
-        stored_state_transition,
-    ) {
-        panic!("Failed to put proof data in the ledger db: {}", e);
-    }
-    Ok(())
-}
-
-pub(crate) fn save_commitments<DB>(
-    ledger_db: DB,
-    sequencer_commitments: &[SequencerCommitment],
-    l1_height: u64,
-) where
-    DB: ProverLedgerOps,
-{
-    for sequencer_commitment in sequencer_commitments.iter() {
-        // Save commitments on prover ledger db
-        ledger_db
-            .update_commitments_on_da_slot(l1_height, sequencer_commitment.clone())
-            .unwrap();
-
-        let l2_start_height = sequencer_commitment.l2_start_block_number;
-        let l2_end_height = sequencer_commitment.l2_end_block_number;
-        for i in l2_start_height..=l2_end_height {
-            ledger_db
-                .put_soft_confirmation_status(BatchNumber(i), SoftConfirmationStatus::Proven)
-                .unwrap_or_else(|_| {
-                    panic!(
-                        "Failed to put soft confirmation status in the ledger db {}",
-                        i
-                    )
-                });
-        }
-    }
 }
